workloads/blogbench: Fix shellcheck warnings

+++ workloads/blogbench/blogbench.sh
@@ -1,14 +1,13 @@
 #!/bin/bash
 
-WORKING_DIR=$PWD/`dirname $0`
+WORKING_DIR=$PWD/$(dirname "$0")
 RUN_DIR="run_dir"
 
-pushd $WORKING_DIR
+pushd "$WORKING_DIR"
 
 uname -a
 
-which blogbench
-if [ $? != 0 ]
+if ! which blogbench
 then
 	echo "[error] blogbench not installed"
 	popd
@@ -19,8 +18,8 @@ if [ ! -d $RUN_DIR ]
 then
 	mkdir $RUN_DIR
 fi
-rm -fr $RUN_DIR/*
+rm -fr "${RUN_DIR:?}"/*
 
-blogbench -d $WORKING_DIR/$RUN_DIR
+blogbench -d "$WORKING_DIR/$RUN_DIR"
 
 popd
